+++ CHANGELOG.md
@@ -7,7 +7,6 @@
 - Add modes `pack` (default) and `next` to Grid component's `masonry` option
 - Add `parallax-start`, `parallax-end`, `parallax-justify` option to Grid component
 - Add `close-on-scroll` option to Drop and Dropdown component
-- Add Navbar height as offset to Scroll component
 - Add transparent Navbar gets its color modifier from underlying section 
 - Add support for `<td>` and `<tr>` elements to `fragment()` function
 - Add 3XL size modifier to Heading component
@@ -15,6 +14,7 @@
 
 ### Change
 
+- Scroll component scrolls beneath elements covering its target 
 - The `closest()` function no longer accepts an array of elements
 - Reset fieldset min-width